Giant clusterfuck change that seems to work
Tests still pass but, um, probably want to refactor?

This is because entire debt module was inadvertently
dependent on rubycritic being, instead of just static
analysis component.  This seems like a better separation
of concerns, but a) it may be way slower/not equivalent,
though b) it should be easier to parallelize.  Maybe add
concurrent-ruby and take a pass at that?

Also raise debt ceiling again, because, debt...